Fix complementary palette not generated
JIRA: BB-2814
